% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{coerceDB1list}
\alias{coerceDB1list}
\alias{coerce,filehashDB1,list-method}
\title{Coerce a filehash database}
\arguments{
\item{from}{a filehashDB1 database object}
}
\description{
Coerce a filehashDB1 database to a list object
}
