% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{coercelist}
\alias{coercelist}
\alias{coerce,filehash,list-method}
\title{Coerce a filehash database}
\arguments{
\item{from}{a filehash database object}
}
\description{
Coerce a filehash database to a list object
}
