% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash.R
\name{db2env}
\alias{db2env}
\title{Load Active Bindings}
\usage{
db2env(db)
}
\arguments{
\item{db}{filehash database object}
}
\value{
environment containing database keys
}
\description{
Load active bindings into an environment and return the environment
}
\details{
\code{db2env} loads the entire database \code{db} into an 
environment via calls to \code{makeActiveBinding}.  Therefore, the data 
themselves are not stored in the environment, but a function pointing to 
the data in the database is stored.  When an element of the environment is 
accessed, the function is called to retrieve the data from the database.  
If the data in the database is changed, the changes will be reflected in the 
environment.

The use of \code{makeActiveBinding} in \code{db2env} and 
\code{dbLoad} allows for potentially large databases to, at least 
conceptually, be used in R, as long as you don't need simultaneous access 
to all of the   elements in the database.
}
\seealso{
\code{\link{dbLoad}}, \code{\link{dbLazyLoad}}
}
