% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash.R
\name{dbLazyLoad}
\alias{dbLazyLoad}
\title{Lazy Load a Database}
\usage{
dbLazyLoad(db, ...)
}
\arguments{
\item{db}{a filehash database object}
}
\value{
a character vector is returned (invisibly) containing the keys 
associated with the values loaded into the environment.
}
\description{
Lazy load a filehash database
}
\details{
\code{dbLazyLoad} loads objects in the database directly into the
environment specified, like \code{load} does except with promises. 
\code{dbLazyLoad} takes a second argument \code{env}, which is an 
environment, and the default for \code{env} is \code{parent.frame()}.

With \code{dbLazyLoad} database objects are "lazy-loaded" into 
the environment. Promises to load the objects are created in the environment 
specified by \code{env}.  Upon first access, those objects are copied into 
the environment and will from then on reside in memory.  Changes to the
database will not be reflected in the object residing in the environment 
after first access.  Conversely, changes to the object in the environment 
will not be reflected in the database.  This type of loading is useful for 
read-only databases.
}
