% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash.R
\name{dbLoad}
\alias{dbLoad}
\title{Load a Database}
\usage{
dbLoad(db, ...)
}
\arguments{
\item{db}{filehash database object}
}
\value{
a character vector containing the names of loaded objects
}
\description{
Load entire database into an environment
}
\details{
\code{dbLoad} loads objects in the database directly into the 
environment specified, like \code{load} does except with active bindings. 
\code{dbLoad} takes a second argument \code{env}, which is an 
environment, and the default for \code{env} is \code{parent.frame()}.

The use of \code{makeActiveBinding} in \code{db2env} and 
\code{dbLoad} allows for potentially large databases to, at least 
conceptually, be used in R, as long as you don't need simultaneous access to 
all of the elements in the database.
}
