% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash-DB1.R
\docType{class}
\name{filehashDB1-class}
\alias{filehashDB1-class}
\title{Filehash DB1 Class}
\description{
An implementation of filehash databases using a single large file
}
\section{Slots}{

\describe{
\item{\code{datafile}}{full path to the database file (filehashDB1 only)}

\item{\code{meta}}{list containing an environment for database metadata (filehashDB1 only)}
}}

