% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash-RDS.R
\docType{class}
\name{filehashRDS-class}
\alias{filehashRDS-class}
\title{Filehash RDS Class}
\description{
An implementation of filehash databases using diretories and separate files
}
\section{Slots}{

\describe{
\item{\code{dir}}{Directory where files are stored (filehashRDS only)}
}}

