% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash.R
\name{registerFormatDB}
\alias{registerFormatDB}
\title{Register Database Format}
\usage{
registerFormatDB(name, funlist)
}
\arguments{
\item{name}{character, name of database format}

\item{funlist}{list of functions for creating and initializing a database format}
}
\description{
Register Database Format
}
