% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash.R
\name{filehashFormats}
\alias{filehashFormats}
\title{List and register filehash formats}
\usage{
filehashFormats(...)
}
\arguments{
\item{\dots}{list of functions for registering a new database format}
}
\value{
A list containing information on the available filehash formats
}
\description{
List and register filehash backend database formats.
}
\details{
\code{filehashFormats} can be used to register new filehash backend database formats.  \code{filehashFormats} called with no arguments lists information on available formats
}
