% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash.R
\name{filehashOption}
\alias{filehashOption}
\title{Set Filehash Options}
\usage{
filehashOption(...)
}
\arguments{
\item{\dots}{name-value pairs for options}
}
\value{
\code{filehashOptions} returns a list of current settings for all options.
}
\description{
Set global filehash options
}
\details{
Currently, the only option that can be set is the default database type (\code{defaultType}) which can be "DB1", "RDS" or "DB".
}
